Move some et tests around and delete unused files.
