+++ tools/engine_tool/test/commands/query_command_test.dart
@@ -9,8 +9,8 @@ import 'package:engine_tool/src/commands/command_runner.dart';
 import 'package:logging/logging.dart' as log;
 import 'package:test/test.dart';
 
-import 'src/test_build_configs.dart';
-import 'utils.dart';
+import '../src/test_build_configs.dart';
+import '../src/utils.dart';
 
 void main() {
   List<String> stringsFromLogs(List<log.LogRecord> logs) {
